requirements.txt
python-dotenv==1.0.0
PyQt6==6.4.2
psycopg2==2.9.5
googletrans==3.1.0a0
pytz